# docs/openapi.yaml
openapi: 3.1.0
info:
  title: EV Battery Insight Service API
  version: '1.0.0'
  description: |
    REST interface for ingesting EV telemetry, retrieving insights, and running deterministic simulations.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error constant.
            message:
              type: string
              description: Summary of the failure.
            details:
              description: Optional structured metadata for debugging.
            requestId:
              type: string
              nullable: true
              description: Correlation identifier echoed from the request.
    Alert:
      type: object
      required:
        - title
        - message
        - severity
      properties:
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
    DriverTip:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    RuleImpact:
      type: object
      required:
        - id
        - deduction
      properties:
        id:
          type: string
          description: Stable rule identifier.
        deduction:
          type: number
          format: float
          description: Points removed from the base score.
        metadata:
          type: object
          additionalProperties: true
          description: Context that informed the rule decision.
    NormalizedTelemetry:
      type: object
      required:
        - vehicleId
        - snapshotTimestamp
        - batteryPercentage
        - speedKmph
        - engineOn
        - charging
        - rawTimestamp
      properties:
        vehicleId:
          type: string
        snapshotTimestamp:
          type: string
          format: date-time
        rawTimestamp:
          type: string
          format: date-time
        batteryPercentage:
          type: number
          minimum: 0
          maximum: 100
        speedKmph:
          type: number
          minimum: 0
        engineOn:
          type: boolean
        charging:
          type: boolean
        ambientTemperature:
          type: number
          nullable: true
        odometerKm:
          type: number
          nullable: true
    TelemetryIngestRequest:
      type: object
      required:
        - vehicleId
        - timestamp
        - batteryPercentage
        - speedKmph
        - engineOn
        - charging
      properties:
        vehicleId:
          type: string
        timestamp:
          type: string
          format: date-time
        batteryPercentage:
          type: number
          minimum: 0
          maximum: 100
        speedKmph:
          type: number
          minimum: 0
        engineOn:
          type: boolean
        charging:
          type: boolean
        ambientTemperature:
          type: number
          description: Degrees Celsius.
        odometerKm:
          type: number
    BatteryEvaluation:
      type: object
      required:
        - baseScore
        - score
        - status
        - ruleImpacts
      properties:
        baseScore:
          type: number
        score:
          type: number
        status:
          type: string
          enum: [GOOD, MODERATE, POOR]
        ruleImpacts:
          type: array
          items:
            $ref: '#/components/schemas/RuleImpact'
    TelemetryIngestResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicleId
            - score
            - status
            - baseScore
            - alerts
            - tips
            - ruleImpacts
            - telemetry
          properties:
            vehicleId:
              type: string
            score:
              type: number
            status:
              type: string
              enum: [GOOD, MODERATE, POOR]
            baseScore:
              type: number
            alerts:
              type: array
              items:
                $ref: '#/components/schemas/Alert'
            tips:
              type: array
              items:
                $ref: '#/components/schemas/DriverTip'
            ruleImpacts:
              type: array
              items:
                $ref: '#/components/schemas/RuleImpact'
            telemetry:
              $ref: '#/components/schemas/NormalizedTelemetry'
    InsightHistoryEntry:
      type: object
      required:
        - id
        - createdAt
        - healthScore
        - status
        - alerts
        - tips
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        healthScore:
          type: number
        status:
          type: string
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        tips:
          type: array
          items:
            $ref: '#/components/schemas/DriverTip'
    VehicleInsights:
      type: object
      required:
        - vehicleId
        - score
        - status
        - telemetry
        - evaluation
        - alerts
        - tips
        - history
        - lastSeenAt
      properties:
        vehicleId:
          type: string
        score:
          type: number
        status:
          type: string
          enum: [GOOD, MODERATE, POOR]
        telemetry:
          $ref: '#/components/schemas/NormalizedTelemetry'
        evaluation:
          $ref: '#/components/schemas/BatteryEvaluation'
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        tips:
          type: array
          items:
            $ref: '#/components/schemas/DriverTip'
        history:
          type: array
          items:
            $ref: '#/components/schemas/InsightHistoryEntry'
        lastSeenAt:
          type: string
          format: date-time
    VehicleInsightsResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/VehicleInsights'
    SimulationScenario:
      type: object
      required:
        - name
        - description
        - samples
      properties:
        name:
          type: string
          enum: [urban, highway, mixed]
        description:
          type: string
        samples:
          type: array
          items:
            type: object
            required:
              - minuteOffset
              - batteryPercentage
              - speedKmph
              - engineOn
              - charging
              - ambientTemperature
              - odometerKm
            properties:
              minuteOffset:
                type: number
              batteryPercentage:
                type: number
              speedKmph:
                type: number
              engineOn:
                type: boolean
              charging:
                type: boolean
              ambientTemperature:
                type: number
              odometerKm:
                type: number
    SimulationScenariosResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SimulationScenario'
    SimulationDriveRequest:
      type: object
      required:
        - scenario
      properties:
        scenario:
          type: string
          enum: [urban, highway, mixed]
        vehicleId:
          type: string
          description: Overrides the default simulation vehicle id.
        persist:
          type: boolean
          description: When true, each sample is persisted through the ingestion pipeline.
        baseTimestamp:
          type: string
          format: date-time
          description: Base timestamp for the first sample in the scenario.
    SimulationDriveResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicleId
            - scenario
            - persisted
            - samples
            - before
            - after
            - alerts
            - tips
          properties:
            vehicleId:
              type: string
            scenario:
              $ref: '#/components/schemas/SimulationScenario'
            persisted:
              type: boolean
            samples:
              type: array
              items:
                $ref: '#/components/schemas/NormalizedTelemetry'
            before:
              type: object
              required:
                - batteryPercentage
                - snapshotTimestamp
              properties:
                batteryPercentage:
                  type: number
                snapshotTimestamp:
                  type: string
                  format: date-time
            after:
              type: object
              required:
                - batteryPercentage
                - snapshotTimestamp
                - score
                - status
              properties:
                batteryPercentage:
                  type: number
                snapshotTimestamp:
                  type: string
                  format: date-time
                score:
                  type: number
                status:
                  type: string
                  enum: [GOOD, MODERATE, POOR]
            alerts:
              type: array
              items:
                $ref: '#/components/schemas/Alert'
            tips:
              type: array
              items:
                $ref: '#/components/schemas/DriverTip'
    SmartcarVehicle:
      type: object
      required:
        - id
        - make
        - model
        - year
      properties:
        id:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
    SmartcarConnectResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - url
            - scope
          properties:
            url:
              type: string
              format: uri
            scope:
              type: array
              items:
                type: string
    SmartcarCallbackResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - userId
            - scope
            - expiresAt
            - vehicles
          properties:
            userId:
              type: string
            scope:
              type: array
              items:
                type: string
            expiresAt:
              type: string
              format: date-time
            refreshExpiresAt:
              type: string
              format: date-time
              nullable: true
            vehicles:
              type: array
              items:
                $ref: '#/components/schemas/SmartcarVehicle'
            state:
              type: string
              nullable: true
    SmartcarVehiclesResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicles
          properties:
            vehicles:
              type: array
              items:
                $ref: '#/components/schemas/SmartcarVehicle'
    SmartcarBatteryData:
      type: object
      properties:
        percentRemaining:
          type: number
          nullable: true
        rangeRemainingMeter:
          type: number
          nullable: true
    SmartcarChargeData:
      type: object
      properties:
        isPluggedIn:
          type: boolean
        state:
          type: string
    SmartcarOdometerData:
      type: object
      properties:
        distance:
          type: number
          nullable: true
    SmartcarEngineData:
      type: object
      properties:
        isRunning:
          type: boolean
    SmartcarLocationData:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        time:
          type: string
          format: date-time
    SmartcarVehicleDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicleId
          properties:
            vehicleId:
              type: string
            battery:
              $ref: '#/components/schemas/SmartcarBatteryData'
            charge:
              $ref: '#/components/schemas/SmartcarChargeData'
            odometer:
              $ref: '#/components/schemas/SmartcarOdometerData'
            engine:
              $ref: '#/components/schemas/SmartcarEngineData'
            location:
              $ref: '#/components/schemas/SmartcarLocationData'
    SmartcarVehicleBatteryResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicleId
            - battery
          properties:
            vehicleId:
              type: string
            battery:
              $ref: '#/components/schemas/SmartcarBatteryData'
    SmartcarVehicleChargeResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicleId
            - charge
          properties:
            vehicleId:
              type: string
            charge:
              $ref: '#/components/schemas/SmartcarChargeData'
    SmartcarVehicleOdometerResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicleId
            - odometer
          properties:
            vehicleId:
              type: string
            odometer:
              $ref: '#/components/schemas/SmartcarOdometerData'
    SmartcarVehicleEngineResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicleId
            - engine
          properties:
            vehicleId:
              type: string
            engine:
              $ref: '#/components/schemas/SmartcarEngineData'
    SmartcarVehicleLocationResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - vehicleId
            - location
          properties:
            vehicleId:
              type: string
            location:
              $ref: '#/components/schemas/SmartcarLocationData'
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
    ReadyResponse:
      type: object
      required:
        - status
        - details
      properties:
        status:
          type: string
          enum: [ready]
        details:
          type: object
          required:
            - database
            - appliedMigrations
          properties:
            database:
              type: object
              properties:
                connected:
                  type: boolean
                path:
                  type: string
                lastMigration:
                  type: string
                  nullable: true
              required:
                - connected
                - path
            appliedMigrations:
              type: array
              items:
                type: string
        requestId:
          type: string
          nullable: true
paths:
  /health:
    get:
      summary: Liveness probe
      tags: [Operations]
      responses:
        '200':
          description: Service is running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /ready:
    get:
      summary: Readiness probe
      tags: [Operations]
      responses:
        '200':
          description: Service is ready to receive traffic.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
        '503':
          description: Service dependencies are not ready.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/telemetry:
    post:
      summary: Ingest telemetry snapshot
      tags: [Telemetry]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryIngestRequest'
      responses:
        '201':
          description: Snapshot accepted and scored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryIngestResponse'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/vehicles/{vehicleId}/insights:
    get:
      summary: Retrieve latest vehicle insight
      tags: [Insights]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Insight payload for the requested vehicle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleInsightsResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Vehicle or telemetry not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/simulate/scenarios:
    get:
      summary: List available simulation scenarios
      tags: [Simulation]
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Enumeration of built-in scenarios.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationScenariosResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/simulate/drive:
    post:
      summary: Execute a deterministic drive scenario
      tags: [Simulation]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationDriveRequest'
      responses:
        '201':
          description: Simulation run completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationDriveResponse'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/smartcar/connect:
    get:
      summary: Generate Smartcar Connect URL
      description: Returns a Smartcar Connect authorization URL along with the scopes that will be requested.
      tags: [Smartcar]
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Optional opaque value echoed back during the OAuth callback.
      responses:
        '200':
          description: Connect URL generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartcarConnectResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/smartcar/callback:
    get:
      summary: Handle Smartcar OAuth callback
      description: Exchanges the Smartcar authorization code for tokens, persists them, and returns connected vehicle metadata.
      tags: [Smartcar]
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code returned by Smartcar Connect.
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Optional state parameter echoed from the original Connect request.
      responses:
        '200':
          description: Smartcar user linked successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartcarCallbackResponse'
        '400':
          description: Missing required authorization code or invalid payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/smartcar/vehicles:
    get:
      summary: List Smartcar vehicles
      tags: [Smartcar]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: Specify a Smartcar user id when multiple connections are stored.
      responses:
        '200':
          description: Vehicles linked to the Smartcar user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartcarVehiclesResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/smartcar/vehicles/{vehicleId}/battery:
    get:
      summary: Retrieve Smartcar battery data
      tags: [Smartcar]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: Specify a Smartcar user id when multiple connections are stored.
      responses:
        '200':
          description: Battery state for the vehicle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartcarVehicleBatteryResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/smartcar/vehicles/{vehicleId}/charge:
    get:
      summary: Retrieve Smartcar charge status
      tags: [Smartcar]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Charging status for the vehicle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartcarVehicleChargeResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/smartcar/vehicles/{vehicleId}/odometer:
    get:
      summary: Retrieve Smartcar odometer reading
      tags: [Smartcar]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Odometer reading for the vehicle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartcarVehicleOdometerResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/smartcar/vehicles/{vehicleId}/engine:
    get:
      summary: Retrieve Smartcar engine status
      tags: [Smartcar]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Engine state for the vehicle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartcarVehicleEngineResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/smartcar/vehicles/{vehicleId}/location:
    get:
      summary: Retrieve Smartcar location
      tags: [Smartcar]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Last reported vehicle location.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartcarVehicleLocationResponse'
        '401':
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'